readyを別ファイルに移動
